pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tessera"
version = "0.1.0"
requires-python = ">= 3.8"
description = "A coarse-grained representation for efficient protein structure analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Leonardo V. Castorina"}
]
keywords = ["protein", "fragments", "graph", "bioinformatics", "structural biology", "protein design"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "ampal>=1.5.3",
    "numpy>=1.21.0",
    "pandas>=2.0.0",
    "scipy>=1.9.0",
    "networkx>=2.8.0",
    "matplotlib>=3.5.0",
    "h5py>=3.0.0",
    "tabulate>=0.9.0",
    "biopython>=1.80",
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
]

[project.optional-dependencies]
ml = [
    "lightning>=2.0.0",
    "mlflow>=2.0.0",
    "numba>=0.57.0",
]
viz = [
    "streamlit>=1.20.0",
    "stmol>=0.0.9",
    "py3Dmol>=2.0.0",
    "pyvis>=0.3.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
all = [
    "tessera[ml,viz,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/tessera"
Documentation = "https://github.com/yourusername/tessera#readme"
Repository = "https://github.com/yourusername/tessera"
"Bug Tracker" = "https://github.com/yourusername/tessera/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["tessera*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]